+++ app/models/batch_connect/session.rb
@@ -77,8 +77,14 @@ module BatchConnect
       # @return [Array<Session>] list of sessions
       def all
         db_root.children.select(&:file?).map do |f|
-          new.from_json(f.read)
-        end.map do |s|
+          begin
+            new.from_json(f.read)
+          rescue => e
+            Rails.logger.error("ERROR: Error parsing file '#{f}' --- #{e.class} - #{e.message}")
+            f.delete unless Rails.env.development?
+            nil
+          end
+        end.compact.map do |s|
           s.completed? && s.destroy ? nil : s
         end.compact.sort_by(&:created_at).reverse
       end